// routes/logic/classes/interfaces/ProviderContactInterface.ts
export interface ProviderContactInterface {
    providerName: string;
    providerContactName: string;
    providerContactTitle: string;
    providerContactEmail: string;
    providerContactTelefon: string;
    providerContactPosition: string;
}

export interface ProviderContactDataInterface {
    PROVIDERS_NAME: string;
    PROVIDERS_CONTACT_NAME: string;
    PROVIDERS_CONTACT_TITLE: string;
    PROVIDERS_CONTACT_EMAIL: string;
    PROVIDERS_CONTACT_TELEFON: string;
    PROVIDERS_CONTACT_POSITION: string;
}